Add separate component for NodePopup
