+++ server/src/components/NodePopup.css
@@ -1,0 +1,49 @@
+#node-popUp {
+    display: none;
+    position: absolute;
+    top: 350px;
+    left: 170px;
+    z-index: 299;
+    width: 525px;
+    overflow: auto;
+    background-color: #333333;
+    border-style: solid;
+    border-width: 3px;
+    border-color: darkgrey;
+    padding: 10px;
+    text-align: center;
+    color: white;
+}
+
+div.vis-configuration-wrapper {
+    width: 500px !important;
+}
+
+div.vis-configuration.vis-config-button {
+    display: none !important;
+}
+
+div.vis-configuration.vis-config-option-container {
+    display: none !important;
+}
+
+div.vis-configuration.vis-config-item.vis-config-s2 {
+    left: 10px;
+    background-color: #333333 !important;
+    padding-left: 5px;
+    border-radius: 3px;
+}
+
+div.vis-configuration.vis-config-item.vis-config-s3 {
+    left: 20px;
+    background-color: #333333 !important;
+    padding-left: 5px;
+    border-radius: 3px;
+}
+
+div.vis-configuration.vis-config-item.vis-config-s4 {
+    left: 30px;
+    background-color: #333333 !important;
+    padding-left: 5px;
+    border-radius: 3px;
+}
